edit layout for paper feed
